requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
python-dotenv==1.0.0
pydantic==2.5.1
pydantic[email]==2.5.1
numpy==1.26.2
psycopg2-binary==2.9.9
weaviate-client[embedded-weaviate]==3.25.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.23.5
httpx==0.26.0
pytest-cov==4.1.0
faker==22.5.0
pytest-env==1.1.3

# Dependencies for linguistic and theme analysis
spacy>=3.7.2
textblob==0.15.3
scikit-learn==1.0.2

# Download spaCy English model
# Run: python -m spacy download en_core_web_sm 